+++ src/core/util/bean_utils.h
@@ -24,7 +24,7 @@ public:
         DCHECK(value, "Empty value being built");
         const std::vector<String> elems = value.split(',');
         const array<sp<Builder<T>>> builders = sp<DynamicArray<sp<Builder<T>>>>::make(elems.size());
-        for(uint32_t i = 0; i < elems.size(); i++)
+        for(size_t i = 0; i < elems.size(); i++)
             builders->buf()[i] = factory.getBuilder<Numeric>(elems[i]);
         return builders;
     }
